Controller maodifications and deployment fixes
