+++ README.md
@@ -1,1 +1,4 @@
 # facerecognitionapi
+
+
+heroku git:remote -a <new app name> to fix old application issue
