+++ controllers/register.js
@@ -1,0 +1,36 @@
+const handleRegister = (req, resp, db, bcrypt) => {
+    const { email, name, password } = req.body;
+
+    if (!email || !name || !password) {
+        return resp.status(400).json('Incorrect form submission');
+    }
+    const hash = bcrypt.hashSync(password);
+
+    db.transaction(trx => {
+        trx.insert({
+            hash: hash,
+            email: email
+        })
+            .into('login')
+            .returning('email')
+            .then(loginEmail => {
+                return trx('users')
+                    .returning('*')
+                    .insert({
+                        email: loginEmail[0],
+                        name: name,
+                        joined: new Date()
+                    }).then(user => {
+                        resp.json(user[0])
+                    })
+            })
+            .then(trx.commit)
+            .catch(trx.rollback)
+    }).catch(err => {
+        console.log(err);
+        resp.json("Unable to register user");
+    })
+}
+module.exports = {
+    handleRegister: handleRegister
+};
