+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "script.js",
   "scripts": {
-    "start": "nodemon server.js"
+    "start": "PORT=3000 server.js"
   },
   "author": "Jaiju Jacob",
   "license": "ISC",
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "bcrypt-nodejs": "0.0.3",
+    "clarifai": "^2.9.1",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "knex": "^0.21.0",
